requirements.txt
ccxt>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.7.0
python-dotenv>=0.19.0
python-telegram-bot>=13.0.0
aiohttp>=3.8.0 